requirements.txt
numpy
scikit-learn
scikit-image
matplotlib
pillow
tqdm
shap
torchinfo
umap-learn